requirements.txt
streamlit>=1.40.0
zhipuai>=3.0.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
requests>=2.31.0
